+++ src/mass.hpp
@@ -10,7 +10,7 @@
 #ifndef __MASS_H__
 #define __MASS_H__
 
-#include <assert.h>
+#include <cassert>
 #include <iostream>
 #include <vector>
 
@@ -19,8 +19,8 @@ class Mass {
 public:
     
     Mass();  // default constructor
-    // constructor with parameters:
-    Mass(float, std::vector<float>, std::vector<float>, std::vector<float>);
+    // constructor with parameters
+    Mass(double, std::vector<double>, std::vector<double>, std::vector<double>);
     ~Mass() {}  // destructor
     Mass(const Mass&);  // copy constructor
     Mass& operator=(const Mass&);  // overloaded assignment operator
@@ -33,37 +33,37 @@ public:
     friend std::ostream& operator<<(std::ostream&, const Mass&);
 
     // getter/setter functions:
-    float get_mass() const { return mass; }
-    void set_mass(const float m) { assert(m >= 0); mass = m; }
+    double get_mass() const { return mass; }
+    void set_mass(const double m) { assert(m >= 0); mass = m; }
 
-    std::vector<float> get_pos() const { return pos; }  // returns copy of vec
-    void set_pos(const float, const float, const float);
-    void set_pos(const std::vector<float> p) { pos = p; } 
+    std::vector<double> get_pos() const { return pos; }  // returns copy of vec
+    void set_pos(const double, const double, const double);
+    void set_pos(const std::vector<double> p) { pos = p; }
 
-    std::vector<float> get_vel() const { return vel; }  // returns copy of vec
-    void set_vel(const float, const float, const float);
-    void set_vel(const std::vector<float> v) { vel = v; }
+    std::vector<double> get_vel() const { return vel; }  // returns copy of vec
+    void set_vel(const double, const double, const double);
+    void set_vel(const std::vector<double> v) { vel = v; }
 
-    std::vector<float> get_acc() const { return acc; }  // returns copy of vec
-    void set_acc(const float, const float, const float);
-    void set_acc(const std::vector<float> a) { acc = a; } 
+    std::vector<double> get_acc() const { return acc; }  // returns copy of vec
+    void set_acc(const double, const double, const double);
+    void set_acc(const std::vector<double> a) { acc = a; }
 
     // additional getters for position components for shorthand
-    float x() const { return pos[0]; }
-    float y() const { return pos[1]; }
-    float z() const { return pos[2]; }
+    double x() const { return pos[0]; }
+    double y() const { return pos[1]; }
+    double z() const { return pos[2]; }
 
     // function to free unused memory in member vectors
     void shrink_vectors();
 
 private:
 
-    float mass;  // mass in kg
+    double mass;  // mass in kg
 
     // position, velocity, acceleration vectors:
-    std::vector<float> pos;  // position in meters
-    std::vector<float> vel;  // velocity in meters/second
-    std::vector<float> acc;  // acceleration in meters/(second^2)
+    std::vector<double> pos;  // position in meters
+    std::vector<double> vel;  // velocity in meters/second
+    std::vector<double> acc;  // acceleration in meters/(second^2)
 
 };
